+++ .eslintignore
@@ -1,3 +1,2 @@
 coverage/
 dist/
-Gruntfile.js
