+++ .npmignore
@@ -1,3 +1,4 @@
+.nyc_output/
 coverage/
 images/
 test/
@@ -6,5 +7,4 @@ AUTHORS.md
 bower.json
 CHANGES.md
 CONTRIBUTING.md
-Gruntfile.js
 README.md
