+++ .travis.yml
@@ -5,8 +5,8 @@ node_js:
   - "10"
   - "11"
 script:
-  - npm run ci
+  - npm run build
 after_success:
-  - npm run report-coverage
+  - npm run coverage
 notifications:
   email: false
