+++ package.json
@@ -23,26 +23,30 @@
   "devDependencies": {
     "chai": "^4.2.0",
     "codecov": "^3.1.0",
+    "eslint": "^5.9.0",
     "eslint-config-notninja": "^0.4.0",
-    "grunt": "^1.0.3",
-    "grunt-cli": "^1.3.2",
-    "grunt-contrib-clean": "^2.0.0",
-    "grunt-eslint": "^21.0.0",
-    "grunt-mocha-istanbul": "^5.0.2",
-    "grunt-rollup": "^9.0.0",
-    "istanbul": "^0.4.5",
-    "load-grunt-tasks": "^4.0.0",
     "mocha": "^5.2.0",
-    "rollup-plugin-commonjs": "^8.4.1",
+    "npm-run-all": "^4.1.3",
+    "nyc": "^13.1.0",
+    "rimraf": "^2.6.2",
+    "rollup": "^0.67.3",
+    "rollup-plugin-commonjs": "^9.2.0",
     "rollup-plugin-node-resolve": "^3.4.0",
     "rollup-plugin-uglify": "^6.0.0"
   },
-  "main": "src/pollock.js",
+  "main": "index.js",
   "browser": "dist/pollock.js",
   "scripts": {
-    "build": "grunt build",
-    "ci": "grunt ci",
-    "report-coverage": "codecov",
-    "test": "grunt test"
+    "build": "npm-run-all lint compile test",
+    "clean": "npm-run-all clean:*",
+    "clean:coverage": "rimraf coverage",
+    "clean:dist": "rimraf dist",
+    "compile": "npm run clean:dist && rollup --config",
+    "coverage": "nyc report && codecov",
+    "lint": "npm-run-all lint:*",
+    "lint:src": "eslint \"src/**/*.js\"",
+    "lint:test": "eslint \"test/**/*.js\"",
+    "test": "npm run clean:coverage && nyc mocha -R list \"test/**/*.spec.js\"",
+    "posttest": "nyc check-coverage"
   }
 }
